ai-engine/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.4

# LLM
anthropic==0.28.0
openai==1.35.3
langchain==0.2.5
langchain-anthropic==0.1.15
langchain-openai==0.1.14
langchain-community==0.2.5

# Embeddings & Vector DB
sentence-transformers==3.0.1
chromadb==0.5.3

# Code processing
pygments==2.18.0
tiktoken==0.7.0
gitpython==3.1.43

# Utils
aiofiles==23.2.1
tenacity==8.3.0
